requirements.txt
langchain==0.2.16
langchain-core==0.2.38
langchain-google-genai==1.0.10
langchain-openai==0.1.22
Pillow==10.4.0